[freeroom.android] fixing correctly issue #3 (refreshment issue)
The fix in 5d687e87dd8446adafd7c1840ca3b8aff3bd7cd4 was not efficient,
this one should be better.
The controller stores the request while we are changing activities, and
the new activity ask the controller to send the query to server AFTER
the activity is completely launched.